Did the following: - Removed all custom Q objects.  Updated (I hope) all queries that used them, to use new Django features instead. - Updated ESPUser to deal with the fact that Django now actually wants to treat it as a sub-model of User, since it supports properly subclassing Models now.
Still to do:
- Deal with all the maxlength/max_length deprecation warnings
- ProgramModuleObj is also a model subclass; it needs to be updated accordingly




git-svn-id: http://esp.mit.edu/code/esp-project/branches/django-HEAD-transition@1587 9f9a36c0-4815-0410-8ed4-a58642b8ae3a

+++ esp/esp/accounting_core/models.py
@@ -55,9 +55,9 @@ class LineItemTypeManager(ProcedureManager):
 class LineItemType(models.Model):
     """ A set of default values for a line item """
     text = models.TextField() # description of line item
-    amount = models.FloatField(help_text='This should be negative for student costs charged to a program.', max_digits=9, decimal_places=2, default=0.0) # default amount
+    amount = models.DecimalField(help_text='This should be negative for student costs charged to a program.', max_digits=9, decimal_places=2, default=0.0) # default amount
     anchor = AjaxForeignKey(DataTree,related_name='accounting_lineitemtype',null=True) # account to post the line item
-    finaid_amount = models.FloatField(max_digits=9, decimal_places=2, default=0.0) # amount after financial aid
+    finaid_amount = models.DecimalField(max_digits=9, decimal_places=2, default=0.0) # amount after financial aid
     finaid_anchor = AjaxForeignKey(DataTree,null=True,related_name='accounting_finaiditemtype')
     
     def negative_amount(self):
@@ -81,7 +81,7 @@ class Balance(models.Model):
     anchor = AjaxForeignKey(DataTree, related_name='balance')
     user = AjaxForeignKey(User, related_name='balance')
     timestamp = models.DateTimeField()
-    amount = models.FloatField(max_digits=16, decimal_places=2)
+    amount = models.DecimalField(max_digits=16, decimal_places=2)
     past = models.ForeignKey('self', null=True)
 
     def negative_amount(self):
@@ -251,7 +251,7 @@ class LineItem(models.Model):
     transaction = models.ForeignKey(Transaction)
     user = AjaxForeignKey(User,related_name='accounting_lineitem')
     anchor = AjaxForeignKey(DataTree,related_name='accounting_lineitem')
-    amount = models.FloatField(max_digits=9, decimal_places=2)
+    amount = models.DecimalField(max_digits=9, decimal_places=2)
     text = models.TextField()
     li_type = models.ForeignKey(LineItemType)
     posted_to = models.ForeignKey(Balance, null=True)
